fix: change floats to doubles

+++ services/wastage/recommendation/kubernetes_pod.go
@@ -62,11 +62,11 @@ func (s *Service) KubernetesPodRecommendation(
 				s.logger.Error("invalid CPUBreathingRoom value", zap.String("value", v.Value))
 				return nil, echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("invalid CpuBreathingRoom value: %s", *v))
 			}
-			recommended.CpuLimit = float32(calculateHeadroom(float64(recommended.CpuLimit), vPercent))
+			recommended.CpuLimit = calculateHeadroom(recommended.CpuLimit, vPercent)
 			if recommended.CpuLimit < 0.1 {
 				recommended.CpuLimit = 0.1
 			}
-			recommended.CpuRequest = float32(calculateHeadroom(float64(recommended.CpuRequest), vPercent))
+			recommended.CpuRequest = calculateHeadroom(recommended.CpuRequest, vPercent)
 			if recommended.CpuRequest < 0.1 {
 				recommended.CpuRequest = 0.1
 			}
@@ -78,24 +78,24 @@ func (s *Service) KubernetesPodRecommendation(
 				s.logger.Error("invalid MemoryBreathingRoom value", zap.String("value", v.Value))
 				return nil, echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("invalid MemoryBreathingRoom value: %s", *v))
 			}
-			recommended.MemoryLimit = float32(calculateHeadroom(float64(recommended.MemoryLimit), vPercent))
+			recommended.MemoryLimit = calculateHeadroom(recommended.MemoryLimit, vPercent)
 			if recommended.MemoryLimit == 0 {
 				recommended.MemoryLimit = 100 * (1024 * 1024)
 			}
-			recommended.MemoryRequest = float32(calculateHeadroom(float64(recommended.MemoryRequest), vPercent))
+			recommended.MemoryRequest = calculateHeadroom(recommended.MemoryRequest, vPercent)
 			if recommended.MemoryRequest == 0 {
 				recommended.MemoryRequest = 100 * (1024 * 1024)
 			}
 		}
 
-		var usageMemoryTrimmedMean, usageMemoryMax, usageCpuTrimmedMean, usageCpuMax *wrappers.FloatValue
+		var usageMemoryTrimmedMean, usageMemoryMax, usageCpuTrimmedMean, usageCpuMax *wrappers.DoubleValue
 		if len(metrics[container.Name].Cpu) > 0 {
-			usageCpuTrimmedMean = wrapperspb.Float(cpuTrimmedMean)
-			usageCpuMax = wrapperspb.Float(cpuMax)
+			usageCpuTrimmedMean = wrapperspb.Double(cpuTrimmedMean)
+			usageCpuMax = wrapperspb.Double(cpuMax)
 		}
 		if len(metrics[container.Name].Memory) > 0 {
-			usageMemoryTrimmedMean = wrapperspb.Float(memoryTrimmedMean)
-			usageMemoryMax = wrapperspb.Float(memoryMax)
+			usageMemoryTrimmedMean = wrapperspb.Double(memoryTrimmedMean)
+			usageMemoryMax = wrapperspb.Double(memoryMax)
 		}
 
 		containersRightsizing = append(containersRightsizing, &pb.KubernetesContainerRightsizingRecommendation{
@@ -120,11 +120,11 @@ func (s *Service) KubernetesPodRecommendation(
 	}, nil
 }
 
-func getMetricMax(data map[string]float32) float32 {
+func getMetricMax(data map[string]float64) float64 {
 	if len(data) == 0 {
 		return 0
 	}
-	dMax := float32(0)
+	dMax := float64(0)
 	for _, v := range data {
 		if v > dMax {
 			dMax = v
@@ -133,24 +133,24 @@ func getMetricMax(data map[string]float32) float32 {
 	return dMax
 }
 
-func getTrimmedMean(data map[string]float32, trimPercentage float32) float32 {
+func getTrimmedMean(data map[string]float64, trimPercentage float64) float64 {
 	if len(data) == 0 {
 		return 0
 	}
 
 	values := make([]float64, 0, len(data))
 	for _, v := range data {
-		values = append(values, float64(v))
+		values = append(values, v)
 	}
 
 	sort.Float64s(values)
 
-	numToTrim := int(trimPercentage * float32(len(data)) / 2)
+	numToTrim := int(trimPercentage * float64(len(data)) / 2)
 	trimmedValues := values[numToTrim : len(values)-numToTrim]
 
 	var sum float64
 	for _, v := range trimmedValues {
 		sum += v
 	}
-	return float32(sum) / float32(len(trimmedValues))
+	return float64(sum) / float64(len(trimmedValues))
 }
